+++ lizard_portal/templates/portals/beheer.js
@@ -88,6 +88,13 @@
                      disabled: false,
                      {% endif %}
                      handler: function() { window.open('/task/') }
+                },
+		{
+                     text: 'Upload configuratie file',
+                     {% if user.is_authenticated %}
+                     disabled: false,
+                     {% endif %}
+                     handler: function() { window.open('/portal/uploadfile/') }
                 }
             ]
 		}]